+++ cli/lib.rs
@@ -23,7 +23,6 @@ extern crate url;
 
 mod checksum;
 pub mod colors;
-pub mod compilers;
 pub mod deno_dir;
 pub mod diagnostics;
 mod disk_cache;
@@ -57,6 +56,7 @@ mod swc_util;
 mod test_runner;
 pub mod test_util;
 mod tokio_util;
+mod tsc;
 mod upgrade;
 pub mod version;
 mod web_worker;
@@ -66,7 +66,6 @@ pub use dprint_plugin_typescript::swc_common;
 pub use dprint_plugin_typescript::swc_ecma_ast;
 pub use dprint_plugin_typescript::swc_ecma_parser;
 
-use crate::compilers::TargetLib;
 use crate::doc::parser::DocFileLoader;
 use crate::file_fetcher::SourceFile;
 use crate::file_fetcher::SourceFileFetcher;
@@ -76,6 +75,7 @@ use crate::op_error::OpError;
 use crate::ops::io::get_stdio;
 use crate::state::DebugType;
 use crate::state::State;
+use crate::tsc::TargetLib;
 use crate::worker::MainWorker;
 use deno_core::v8_set_flags;
 use deno_core::ErrBox;
